[WIP] Add draft of PartitionTimingData class...
 - Add draft of `PartitionTimingData`, currently including the following
   two device vectors:
  - `block_keys`
  - `net_driver_block_key`

Notes
=====

The basic idea here is to pack all data for connections related to
blocks that have not had an arrival time resolved to the start of the
vectors.  This is similar to filtering a `pandas.DataFrame` using fancy
iterator queries, except that we are simply reordering the data, and all
data is still present in the vectors.

TODO
====

 - Add all columns of the `arrival_connections` data frame to the
   `PartitionTimingData` class.