Fix departure times calculation.
Prior to this commit, the source type was being overridden to
`CONNECTION_DRIVER` in the `prepare_device_timing_data` function
regardless of the value of the `source` argument.  The effect of this
bug is that the arrival times are _always_ computed even when the
departure times are expected in the case where `CONNECTION_SINK` is
specified using the `source` argument.

This commit preserves the source type specified by the `source`
argument.